community page to TSX
